refactor(slurm): unify SLURM client naming under slurm/ package (#160)
Phase 4 of #156. Establishes the canonical naming for the SLURM
execution transport layer and resolves the three-way mismatch
between ``Slurm`` (bare), ``SSHSlurmClient`` (verbose), and
``SlurmClientProtocol`` (redundant suffix).

Canonical names under ``src/srunx/slurm/``:
- ``Client`` (Protocol)          ← ``SlurmClientProtocol``
- ``LocalClient``                ← ``Slurm``
- ``JobOperations`` (Protocol)   ← ``JobOperationsProtocol``
- ``WorkflowJobExecutor``        ← ``WorkflowJobExecutorProtocol``
- ``JobSnapshot``                ← ``JobStatusInfo``
- ``LogChunk`` (unchanged)

Moves / creations:
- ``src/srunx/slurm/__init__.py`` — canonical public surface with
  lazy ``__getattr__`` for ``LocalClient`` (PEP 562) to avoid an
  import cycle against ``srunx.utils``.
- ``src/srunx/slurm/local.py`` ← content of ``src/srunx/client.py``
  with class ``Slurm`` renamed to ``LocalClient``.
- ``src/srunx/slurm/protocols.py`` ← content of
  ``src/srunx/client_protocol.py`` with renamed classes.
- ``src/srunx/slurm/parsing.py`` — new; hosts ``GPU_TRES_RE`` and
  re-exports the datetime / duration parsers.

Shims (backward-compat aliases, removed post-migration):
- ``src/srunx/client.py``: ``Slurm = LocalClient`` + re-exports.
- ``src/srunx/client_protocol.py``: old ``*Protocol`` names +
  ``JobStatusInfo`` alias classes.
- ``src/srunx/utils.py``: ``GPU_TRES_RE`` re-imported from
  ``srunx.slurm.parsing``.

Exception naming consistency (bundles #169):
- ``exceptions.JobNotFound`` → ``JobNotFoundError`` with
  ``JobNotFound = JobNotFoundError`` alias preserved.

``src/srunx/__init__.py`` exports ``LocalClient`` alongside
``Slurm``.

Test updates (monkeypatch targets moved with the real code):
- ``tests/test_client.py``: ``srunx.client.render_*`` /
  ``srunx.client.get_job_status`` patches retargeted to
  ``srunx.slurm.local.*``. ``srunx.client.Slurm.*`` patches on
  methods still work via the class alias.

``tests/architecture/test_import_layers.py`` whitelist keys
updated to track the file move (``client`` → ``slurm``).

Verified: architecture tests pass; 1903 passed in full suite
(2 pre-existing mcp env skips). Two tests are flaky under full-
suite concurrency load but pass in isolation — unrelated to this
change.
